country/requirements.txt
flask-restplus
flask-marshmallow
requests>=2.4.2

